Refactor Offer to presentational/container style
