+++ src/features/offer/OfferContainer.tsx
@@ -5,26 +5,15 @@ import { useLocation, useParams, useHistory } from 'react-router-dom';
 import { Notification } from 'hds-react';
 import { getOperationName } from 'apollo-link';
 
-import styles from './offer.module.scss';
 import LoadingSpinner from '../../common/spinner/LoadingSpinner';
 import { OFFER_QUERY } from './queries';
-import Table, { Column, COLUMN_WIDTH } from '../../common/table/Table';
 import Offer from './Offer';
-import InternalLink from '../../common/internalLink/InternalLink';
 import { OFFER } from './__generated__/OFFER';
-import { BerthData, getOfferData, getAllPiersIdentifiers, getBoat, getHarbor } from './utils';
-import { formatDimension, formatDate } from '../../common/utils/format';
+import { getOfferData, getAllPiersIdentifiers, getBoat, getHarbor } from './utils';
+import { formatDate } from '../../common/utils/format';
 import { CREATE_LEASE_MUTATION } from './mutations';
 import { CREATE_LEASE, CREATE_LEASEVariables as CREATE_LEASE_VARS } from './__generated__/CREATE_LEASE';
-import TableTools from './tableTools/TableTools';
-import BerthDetails from '../../common/berthDetails/BerthDetails';
-import TableFilters from '../../common/tableFilters/TableFilters';
 import { BERTH_APPLICATIONS_QUERY } from '../applicationList/queries';
-import HarborCard from '../../common/harborCard/HarborCard';
-import BoatCard from '../../common/boatCard/BoatCard';
-import Button from '../../common/button/Button';
-
-type ColumnType = Column<BerthData> & { accessor: keyof BerthData };
 
 function useRouterQuery() {
   return new URLSearchParams(useLocation().search);
@@ -45,76 +34,6 @@ const OfferContainer = () => {
     }
   );
   const { t, i18n } = useTranslation();
-
-  const columns: ColumnType[] = [
-    {
-      Cell: ({ row }) => (
-        <Button
-          onClick={() => {
-            createBerthLease({
-              variables: {
-                input: {
-                  applicationId: applicationId || '',
-                  berthId: row.original.berthId,
-                },
-              },
-            });
-
-            history.push('/applications');
-          }}
-          disabled={isSubmitting}
-        >
-          {t('offer.tableCells.select')}
-        </Button>
-      ),
-      Header: t('offer.tableHeaders.selection') || '',
-      accessor: 'berthId',
-      width: COLUMN_WIDTH.S,
-      disableFilters: true,
-      disableSortBy: true,
-    },
-    {
-      Cell: ({ cell }) => <InternalLink to={`/harbors/${cell.row.original.harborId}}`}>{cell.value}</InternalLink>,
-      Header: t('offer.tableHeaders.harbor') || '',
-      accessor: 'harbor',
-      width: COLUMN_WIDTH.XL,
-    },
-    {
-      Header: t('offer.tableHeaders.pier') || '',
-      accessor: 'pier',
-      filter: 'exactText',
-      width: COLUMN_WIDTH.S,
-    },
-    {
-      Header: t('offer.tableHeaders.berth') || '',
-      accessor: 'berth',
-      width: COLUMN_WIDTH.XS,
-    },
-    {
-      Cell: ({ cell }) => formatDimension(cell.value, i18n.language),
-      Header: t('offer.tableHeaders.width') || '',
-      accessor: 'width',
-      width: COLUMN_WIDTH.XS,
-    },
-    {
-      Cell: ({ cell }) => formatDimension(cell.value, i18n.language),
-      Header: t('offer.tableHeaders.length') || '',
-      accessor: 'length',
-      width: COLUMN_WIDTH.XS,
-    },
-    {
-      Cell: ({ cell }) => formatDimension(cell.value, i18n.language),
-      Header: t('offer.tableHeaders.draught') || '',
-      accessor: 'draught',
-      width: COLUMN_WIDTH.XS,
-    },
-    {
-      Cell: ({ cell }) => t([`common.mooringTypes.${cell.value}`, cell.value]),
-      Header: t('offer.tableHeaders.mooringType') || '',
-      accessor: 'mooringType',
-      width: COLUMN_WIDTH.S,
-    },
-  ];
 
   if (loading) return <LoadingSpinner isLoading={loading} />;
   if (!data?.berthApplication)
@@ -145,35 +64,30 @@ const OfferContainer = () => {
   const harbor = getHarbor(data);
   const boat = getBoat(data);
 
+  const handleClickSelect = (berthId: string) => {
+    createBerthLease({
+      variables: {
+        input: {
+          applicationId: applicationId || '',
+          berthId,
+        },
+      },
+    }).then(() => history.push('/applications'));
+  };
+
   return (
-    <Offer>
-      {harbor && <HarborCard {...harbor} className={styles.card} />}
-      {boat && <BoatCard boat={boat} />}
-      <Table
-        data={tableData}
-        columns={columns}
-        renderSubComponent={(row) => {
-          const { properties, leases, comment } = row.original;
-          return <BerthDetails leases={leases} comment={comment} {...properties} />;
-        }}
-        renderMainHeader={(props) => (
-          <TableFilters
-            activeFilters={props.state.filters.map((filter) => filter.value)}
-            filters={piersIdentifiers}
-            handleSetFilter={(filter) => props.setFilter('pier', filter)}
-            filterPrefix={t('offer.tableHeaders.pierFilterBtn')}
-          />
-        )}
-        renderTableToolsTop={() => (
-          <TableTools
-            applicationDate={applicationDate}
-            applicationType={applicationType}
-            applicationStatus={applicationStatus}
-            handleReturn={handleReturn}
-          />
-        )}
-      />
-    </Offer>
+    <Offer
+      applicationDate={applicationDate}
+      applicationStatus={applicationStatus}
+      applicationType={applicationType}
+      boat={boat}
+      handleClickSelect={handleClickSelect}
+      handleReturn={handleReturn}
+      harbor={harbor}
+      isSubmitting={isSubmitting}
+      piersIdentifiers={piersIdentifiers}
+      tableData={tableData}
+    />
   );
 };
